refactor: 60-remove types & fix paths
