+++ src/app/app-state.ts
@@ -1,8 +1,8 @@
 import type { AppState } from './types';
 
-import { Page } from '../utils/router/types';
+import { Page } from './constants';
 
 export const appState: AppState = {
   currentPage: Page.main,
-  isLogined: false,
+  isLogined: true,
 };
