+++ src/app/constants.ts
@@ -1,3 +1,5 @@
+import type { PageType } from './types';
+
 export const CLIENT_ID = import.meta.env.VITE_CLIENT_ID;
 export const CLIENT_SECRET = import.meta.env.VITE_CLIENT_SECRET;
 export const API_URL = import.meta.env.VITE_API_URL;
@@ -5,3 +7,14 @@ export const AUTH_URL = import.meta.env.VITE_AUTH_URL;
 export const SCOPES = import.meta.env.VITE_SCOPES;
 export const PROJECT_KEY = import.meta.env.VITE_PROJECT_KEY;
 
+export const Page: Record<PageType, PageType> = {
+  about: 'about',
+  basket: 'basket',
+  catalog: 'catalog',
+  error: 'error',
+  login: 'login',
+  main: 'main',
+  product: 'product',
+  profile: 'profile',
+  registration: 'registration',
+};
