+++ src/utils/router/render-page.ts
@@ -1,18 +1,18 @@
 import type { PageType } from '../../app/types';
 
 import { appState } from '../../app/app-state';
+import { Page } from '../../app/constants';
 import { Header } from '../../components/header';
 import { toggleClassesOnRedirect } from '../../helpers/toggle-classes-on-redirect';
 import { ErrorPage } from '../../pages/error';
 import { Main } from '../../pages/main';
 import { RegistrationPage } from '../../pages/registration/registration';
-import { undercunstruction } from '../../pages/undercunstruction';
-import { Page } from './types';
+import { underconstruction } from '../../pages/underconstruction';
 
 export function renderPage(page: PageType): void {
   toggleClassesOnRedirect(appState.isLogined, page);
 
-  document.body.textContent = '';
+  document.body.replaceChildren();
   Header();
   switch (page) {
     case Page.error: {
@@ -24,11 +24,11 @@ export function renderPage(page: PageType): void {
       break;
     }
     case Page.registration: {
-      document.body.append(RegistrationPage());
+      RegistrationPage();
       break;
     }
     default: {
-      undercunstruction();
+      underconstruction();
     }
   }
 }
