--- src/utils/router/types.ts
@@ -1,11 +1,0 @@
-export const Page = {
-  about: 'about',
-  basket: 'basket',
-  catalog: 'catalog',
-  error: 'error',
-  login: 'login',
-  main: 'main',
-  product: 'product',
-  profile: 'profile',
-  registration: 'registration',
-} as const;
